Add/use EventSandbox component
